Move ProductionQueueFromSelection into RA and generalize.
